+++ CHANGELOG.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## 0.1.2 - unreleased
+
+### Added
+
+* setting for combi matrix to be computable by default
 
 ### Fixed
 
 * group and filter function in issue queries
+* deface errors
 
 ## 0.1.1 - 2021-12-03